Clean request subscriptions on reply
In order to achieve that the request result is first routed to the Nats
State update function which does the cleanup then resend it tagged in a
Task.succeed.

Doing that (resend a message not to self) requires the update function
to return Cmd msg instead of Cmd Msg, hence we have to store a tagger in
the state. Which implies the tagger is no longer required as a 'listen'
argument.